dev-requirements.txt
agate<1.6.2,>=1.6
cffi<1.15,>=1.9
hologram>=0.0.14
idna<2.10
isodate<0.7,>=0.6
Jinja2==2.11.3
json-rpc<2,>=1.12
logbook<1.6,>=1.5
mashumaro==2.5
minimal-snowplow-tracker==0.0.2
colorama<0.4.4,>=0.3.9
pytest-dbt-adapter
urllib3!=1.25.0,!=1.25.1,<1.26,>=1.21.1
typing-extensions<3.8,>=3.7.4
chardet<4,>=3.0.2
scipy!=1.6.1,>=1.5
matplotlib>=3.3
pandas>=1.1
decorator<6,>=5.0.7
pytest-dbt-adapter==0.4.0
waiter
requests<2.24.0,>=2.18.0
wheel
